+++ malachite-q/src/test_util/gaussian_rational/arithmetic/mod.rs
@@ -10,4 +10,5 @@ pub mod add;
 pub mod div;
 pub mod mul;
 pub mod pow;
+pub mod sqrt;
 pub mod square;